feat: implement command-line interface
Height-6 conversion tools were also moved to a separate `height_6` file (still imported in the module `__init__.py` for backwards compatibility).

This was originally implemented by Alexandre Hassan in #37, but I've redone the work in a fresh tree after some refactoring resulting in annoying merge conflicts.

Closes: #36
